requirements.txt
gradio 
requests
google-generativeai